# docs/swagger.yaml
definitions:
...
  github_com_hamidgh01_Go-Blog-API_internal_http_dto.CreateCommentRequest:
    properties:
      content:
        maxLength: 1500
        type: string
      parent_id:
        type: integer
      parent_type:
        enum:
        - post
        - comment
        type: string
    required:
    - content
    - parent_id
    - parent_type
    type: object
...
  github_com_hamidgh01_Go-Blog-API_internal_http_dto.UpdateCommentRequest:
    properties:
      content:
        maxLength: 1500
        type: string
    required:
    - content
    type: object
...
  github_com_hamidgh01_Go-Blog-API_internal_http_dto.UpdateListPrivacyRequest:
    properties:
      is_private:
        type: boolean
    required:
    - is_private
    type: object
...
  github_com_hamidgh01_Go-Blog-API_internal_http_helpers.standardResponse:
    properties:
      data:
        description: presents only for successful response (`success=true`)
      details:
        description: presents only for error response (`success=false`)
      message:
        description: a message that gives more information about response
        type: string
      success:
        description: '`true` for successful response and `false` for error response'
        type: boolean
    type: object
info:
  contact:
    email: hamidghahremani2001@gmail.com
    name: Hamid Ghahremani
    url: https://github.com/hamidgh01
  description: Go powered RESTful Blog API.
  license:
    name: MIT
    url: https://mit-license.org/
  title: Go Blog API
  version: "1.0"
paths:
  /v1/comments:
    post:
      consumes:
      - application/json
      description: Create a new comment
      parameters:
      - description: Data to create new comment
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_dto.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_helpers.standardResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_dto.CommentDetails'
                details:
                  type: object
              type: object
        "400":
          description: validation error
        "422":
          description: invalid json format or json fields with invalid value type
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Create Comment
      tags:
      - comments
  /v1/comments/{id}:
    delete:
      description: Delete a comment (**only admin access**)
      parameters:
      - description: Unique Comment ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Success
        "400":
          description: invalid path parameter
        "404":
          description: comment not found
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Delete Comment
      tags:
      - comments
    get:
      description: Get a comment by unique id
      parameters:
      - description: Unique Comment ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_helpers.standardResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_dto.CommentDetails'
                details:
                  type: object
              type: object
        "400":
          description: invalid path parameter
        "404":
          description: comment not found
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Get Comment By ID
      tags:
      - comments
    put:
      consumes:
      - application/json
      description: Update a comment
      parameters:
      - description: Unique Comment ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Data to update a comment
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_dto.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Success
        "400":
          description: validation error | invalid path parameter
        "404":
          description: comment not found
        "422":
          description: invalid json format or json fields with invalid value type
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Update Comment
      tags:
      - comments
  /v1/comments/{id}/delete:
    patch:
      description: Delete a comment at user request (change status to "deleted")
      parameters:
      - description: Unique Comment ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Success
        "400":
          description: invalid path parameter
        "404":
          description: comment not found
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Delete Comment At User Request
      tags:
      - comments
  /v1/comments/{id}/hide:
    patch:
      description: Hide a comment (change status to "hidden") (**only admin access**)
      parameters:
      - description: Unique Comment ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Success
        "400":
          description: invalid path parameter
        "404":
          description: comment not found
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Hide Comment
      tags:
      - comments
  /v1/comments/{id}/replies:
    get:
      description: Get a list of comment's replies with pagination
      parameters:
      - description: Unique Comment ID to use as FK
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - default: 1
        description: 'Page Number (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: 'Items per Page (default: 10)'
        in: query
        maximum: 100
        minimum: 1
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_helpers.standardResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_generics.PagedList-github_com_hamidgh01_Go-Blog-API_internal_http_dto_CommentList'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_dto.CommentDetails'
                        type: array
                    type: object
                details:
                  type: object
              type: object
        "400":
          description: invalid path parameter | invalid query parameters (for 'size'
            or 'page')
        "404":
          description: there's no any reply for this comment
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Get Comment's Replies
      tags:
      - comments
  /v1/comments/{id}/republish:
    patch:
      description: Republish a comment (change status to "published") (**only admin
        access**)
      parameters:
      - description: Unique Comment ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Success
        "400":
          description: invalid path parameter
        "404":
          description: comment not found
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Republish Comment
      tags:
      - comments
  /v1/links:
    post:
      consumes:
      - application/json
      description: Create a new link
      parameters:
      - description: Data to create new link
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_dto.CreateLinkRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_helpers.standardResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_dto.LinkDetails'
                details:
                  type: object
              type: object
        "400":
          description: validation error
        "422":
          description: invalid json format or json fields with invalid value type
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Create Link
      tags:
      - links
  /v1/links/{id}:
    delete:
      description: Delete a link
      parameters:
      - description: Unique Link ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Success
        "400":
          description: invalid path parameter
        "404":
          description: link not found
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Delete Link
      tags:
      - links
    get:
      description: Get a link by unique id
      parameters:
      - description: Unique Link ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_helpers.standardResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_dto.LinkDetails'
                details:
                  type: object
              type: object
        "400":
          description: invalid path parameter
        "404":
          description: link not found
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Get Link By ID
      tags:
      - links
    put:
      consumes:
      - application/json
      description: Update a link
      parameters:
      - description: Unique Link ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Data to update a link
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_dto.UpdateLinkRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Success
        "400":
          description: validation error | invalid path parameter
        "404":
          description: link not found
        "422":
          description: invalid json format or json fields with invalid value type
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Update Link
      tags:
      - links
  /v1/lists:
    post:
      consumes:
      - application/json
      description: Create a new list
      parameters:
      - description: Data to create new list
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_dto.CreateListRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_helpers.standardResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_dto.ListDetails'
                details:
                  type: object
              type: object
        "400":
          description: validation error
        "422":
          description: invalid json format or json fields with invalid value type
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Create List
      tags:
      - lists
  /v1/lists/{id}:
    delete:
      description: Delete a list
      parameters:
      - description: Unique List ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Success
        "400":
          description: invalid path parameter
        "404":
          description: list not found
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Delete List
      tags:
      - lists
    get:
      description: Get a list by unique id
      parameters:
      - description: Unique List ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_helpers.standardResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_dto.ListDetails'
                details:
                  type: object
              type: object
        "400":
          description: invalid path parameter
        "404":
          description: list not found
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Get List By ID
      tags:
      - lists
    patch:
      consumes:
      - application/json
      description: Update list's privacy
      parameters:
      - description: Unique List ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: new privacy value to update List.IsPrivate
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_dto.UpdateListPrivacyRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Success
        "400":
          description: validation error | invalid path parameter
        "404":
          description: list not found
        "422":
          description: invalid json format or json fields with invalid value type
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Updata List.IsPrivate
      tags:
      - lists
    put:
      consumes:
      - application/json
      description: Update a list
      parameters:
      - description: Unique List ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Data to update a list
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_dto.UpdateListRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Success
        "400":
          description: validation error | invalid path parameter
        "404":
          description: list not found
        "422":
          description: invalid json format or json fields with invalid value type
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Update List
      tags:
      - lists
  /v1/lists/{id}/save-list:
    post:
      description: Save a list
      parameters:
      - description: Unique List ID to Save
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Success
        "400":
          description: invalid path parameter
        "406":
          description: no such list to save (trying to save a list that not exists)
            | list is already saved by this user
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Save List
      tags:
      - lists
  /v1/lists/{id}/saved-posts:
    get:
      description: Get list of saved posts of the list with pagination
      parameters:
      - description: Unique List ID to use as FK
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - default: 1
        description: 'Page Number (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: 'Items per Page (default: 10)'
        in: query
        maximum: 100
        minimum: 1
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_helpers.standardResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_generics.PagedList-github_com_hamidgh01_Go-Blog-API_internal_http_dto_PostsList'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_dto.PostBrief'
                        type: array
                    type: object
                details:
                  type: object
              type: object
        "400":
          description: invalid path parameter | invalid query parameters (for 'size'
            or 'page')
        "404":
          description: there's no any saved-post in this list
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Get List's Saved Posts
      tags:
      - lists
  /v1/lists/{id}/unsave-list:
    delete:
      description: Unsave a list
      parameters:
      - description: Unique List ID to Unsave
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Success
        "400":
          description: invalid path parameter
        "406":
          description: no such list to unsave (trying to unsave a list that not exists)
            | list already isn't saved by this user
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Unsave List
      tags:
      - lists
  /v1/lists/{id}/users-who-saved:
    get:
      description: Get list of users who saved the list with pagination
      parameters:
      - description: Unique List ID to use as FK
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - default: 1
        description: 'Page Number (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: 'Items per Page (default: 10)'
        in: query
        maximum: 100
        minimum: 1
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_helpers.standardResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_generics.PagedList-github_com_hamidgh01_Go-Blog-API_internal_http_dto_UsersList'
                  - properties:
                      items:
                        items:
                          $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_dto.UserBrief'
                        type: array
                    type: object
                details:
                  type: object
              type: object
        "400":
          description: invalid path parameter | invalid query parameters (for 'size'
            or 'page')
        "404":
          description: there's no any user who saves this list
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Get Users Who Saved The List
      tags:
      - lists
  /v1/login:
    post:
      consumes:
      - application/json
      description: Login by identifier (**username** or **email**) and **password**
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_helpers.standardResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_dto.LoginSuccessfulData'
                details:
                  type: object
              type: object
        "400":
          description: validation error
        "401":
          description: invalid credentials
        "403":
          description: suspended user
        "422":
          description: invalid json format or json fields with invalid value type
        "500":
          description: internal server error
      summary: Login
      tags:
      - auth
  /v1/logout:
    get:
      description: Logout a user
      produces:
      - application/json
      responses:
        "200":
          description: Success
        "401":
          description: invalid token | expired token | blacklisted token | not providing
            'refresh_token' cookie
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - auth
  /v1/posts:
    post:
      consumes:
      - application/json
      description: Create a new post
      parameters:
      - description: Data to create new post
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_dto.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_helpers.standardResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_dto.PostDetails'
                details:
                  type: object
              type: object
        "400":
          description: validation error
        "422":
          description: invalid json format or json fields with invalid value type
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Create Post
      tags:
      - posts
  /v1/posts/{id}:
    delete:
      description: Delete a post (**only admin access**)
      parameters:
      - description: Unique Post ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Success
        "400":
          description: invalid path parameter
        "404":
          description: post not found
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Delete Post
      tags:
      - posts
    get:
      description: Get a post by unique id
      parameters:
      - description: Unique Post ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_helpers.standardResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_dto.PostDetails'
                details:
                  type: object
              type: object
        "400":
          description: invalid path parameter
        "404":
          description: post not found
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Get Post By ID
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Update a post
      parameters:
      - description: Unique Post ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Data to update a post
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_hamidgh01_Go-Blog-API_internal_http_dto.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Success
        "400":
          description: validation error | invalid path parameter
        "404":
          description: post not found
        "422":
          description: invalid json format or json fields with invalid value type
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Update Post
      tags:
      - posts
...
  /v1/posts/{id}/delete:
    patch:
      description: Delete a post at user request (change status to "deleted")
      parameters:
      - description: Unique Post ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Success
        "400":
          description: invalid path parameter
        "404":
          description: post not found
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Delete Post At User Request
      tags:
      - posts
...
  /v1/posts/{id}/like:
    post:
      description: Like a post
      parameters:
      - description: Unique Post ID to Like
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Success
        "400":
          description: invalid path parameter
        "406":
          description: no such post to like (trying to like a post that not exists)
            | the target post is already liked by this user
        "500":
          description: internal server error
      security:
      - BearerAuth: []
      summary: Like
      tags:
      - posts
...
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token: "Bearer eyJhbGciOiJIUzI1Ni..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...